factor assignment helpers
